Web: Moved from MediaViewerV2 to MediaViewer and removed old MediaViewer
