+++ packages/common/components/MediaViewer/MediaViewer.props.ts
