--- packages/common/components/MediaViewer/index.js
@@ -1,1 +1,0 @@
-export default from "./MediaViewer";
